An artificial intelligent traffic control system using python and opencv
A project to create an intelligent traffic control system that can
detect traffic density in a traffic situation save the data acquired
from the sensors(camera, sound sensor) and computeand compare to make a
decision as to which roadshould be given a green light for free flow of
traffic. Its both software(Python and OpenCV) and hardware
project(Raspberry Pi, USB camera's, sound sensors).